updated libftprintf.h

+++ libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: kchiang <kchiang@student.42kl.edu.my>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/05 11:40:41 by kchiang           #+#    #+#             */
-/*   Updated: 2025/06/11 12:25:07 by kchiang          ###   ########.fr       */
+/*   Updated: 2025/06/11 12:46:43 by kchiang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ typedef int	(*t_fptr)(va_list, t_spec);
 int		ft_printf(const char *format, ...);
 
 /* Util Function */
-int		ft_putnbrstr(char *str, size_t len, t_spec mod, int is_uphex);
+int		ft_putnbrstr(char *str, int len, t_spec mod, int is_uphex);
 
 /* To be called by t_fptr. */
 /* Functions for their respective conversion specifier. */
